FAC-965:impl:Move insight domain models to insight sub package
