+++ flickit-assessment-core/src/main/java/org/flickit/assessment/core/application/domain/insight/AssessmentInsight.java
@@ -1,4 +1,4 @@
-package org.flickit.assessment.core.application.domain;
+package org.flickit.assessment.core.application.domain.insight;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
@@ -8,10 +8,10 @@ import java.util.UUID;
 
 @Getter
 @RequiredArgsConstructor
-public class SubjectInsight {
+public class AssessmentInsight {
 
+    private final UUID id;
     private final UUID assessmentResultId;
-    private final Long subjectId;
     private final String insight;
     private final LocalDateTime insightTime;
     private final LocalDateTime lastModificationTime;
