+++ flickit-assessment-core/src/test/java/org/flickit/assessment/core/application/service/insight/attribute/CreateAttributeAiInsightServiceTest.java
@@ -7,8 +7,8 @@ import org.flickit.assessment.common.exception.CalculateNotValidException;
 import org.flickit.assessment.common.exception.ResourceNotFoundException;
 import org.flickit.assessment.core.application.domain.AssessmentResult;
 import org.flickit.assessment.core.application.domain.Attribute;
-import org.flickit.assessment.core.application.domain.AttributeInsight;
 import org.flickit.assessment.core.application.domain.MaturityLevel;
+import org.flickit.assessment.core.application.domain.insight.AttributeInsight;
 import org.flickit.assessment.core.application.port.in.insight.attribute.CreateAttributeAiInsightUseCase;
 import org.flickit.assessment.core.application.port.in.insight.attribute.CreateAttributeAiInsightUseCase.Param;
 import org.flickit.assessment.core.application.port.out.assessment.GetAssessmentProgressPort;
