+++ flickit-assessment-core/src/test/java/org/flickit/assessment/core/application/service/insight/subject/CreateSubjectInsightServiceTest.java
@@ -3,7 +3,7 @@ package org.flickit.assessment.core.application.service.insight.subject;
 import org.flickit.assessment.common.application.domain.assessment.AssessmentAccessChecker;
 import org.flickit.assessment.common.exception.AccessDeniedException;
 import org.flickit.assessment.core.application.domain.AssessmentResult;
-import org.flickit.assessment.core.application.domain.SubjectInsight;
+import org.flickit.assessment.core.application.domain.insight.SubjectInsight;
 import org.flickit.assessment.core.application.port.in.insight.subject.CreateSubjectInsightUseCase;
 import org.flickit.assessment.core.application.port.out.assessmentresult.LoadAssessmentResultPort;
 import org.flickit.assessment.core.application.port.out.insight.subject.CreateSubjectInsightPort;
