+++ flickit-assessment-core/src/test/java/org/flickit/assessment/core/application/service/insight/subject/CreateSubjectInsightsHelperTest.java
@@ -3,8 +3,8 @@ package org.flickit.assessment.core.application.service.insight.subject;
 import org.flickit.assessment.common.application.MessageBundle;
 import org.flickit.assessment.core.application.domain.AssessmentResult;
 import org.flickit.assessment.core.application.domain.MaturityLevel;
-import org.flickit.assessment.core.application.domain.SubjectInsight;
 import org.flickit.assessment.core.application.domain.SubjectValue;
+import org.flickit.assessment.core.application.domain.insight.SubjectInsight;
 import org.flickit.assessment.core.application.port.out.maturitylevel.LoadMaturityLevelsPort;
 import org.flickit.assessment.core.application.port.out.subjectvalue.LoadSubjectValuePort;
 import org.flickit.assessment.core.application.service.insight.subject.CreateSubjectInsightsHelper.Param;
